feat(TDP-4285): Cleaning and refactoring
