+++ dataprep-test-api/src/main/java/org/talend/dataprep/qa/step/DatasetStep.java
@@ -4,6 +4,7 @@ import cucumber.api.java8.En;
 import org.junit.Assert;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.talend.dataprep.qa.step.config.DataPrepStep;
 
 /**
  * Step dealing with dataset
