+++ dataprep-test-api/src/main/resources/features/ExportPreparation.feature
@@ -5,4 +5,4 @@ Feature: Export Preparation
     And I create a preparation with name "myFirstPreparation", based on "simpleCSV" dataset
     When I add a step "uppercase" to the column "lastname" of the preparation "myFirstPreparation"
     And I export the preparation "myFirstPreparation" on the dataset "simpleCSV" and export the result in "myResult.csv" temporary file.
-    Then I check that "myResult.csv" temporary file equals "simpleCSV_processed.csv" file
+    Then I check that "myResult.csv" temporary file equals "/data/simpleCSV_processed.csv" file
